[New] Created test cases for pages-admin-form-builder-custom-field-selector-dialog component from enzyme to rtl
