+++ app/javascript/components/pages/admin/form-builder/components/custom-field-selector-dialog/component.spec.js
@@ -1,0 +1,27 @@
+import { fromJS } from "immutable";
+import { mountedComponent, screen } from "test-utils";
+
+import CustomFieldSelectorDialog from "./component";
+
+describe("<CustomFieldSelectorDialog />", () => {
+  const initialState = fromJS({
+    ui: { dialogs: { dialog: "custom_field_selector_dialog", open: true } }
+  });
+
+  beforeEach(() => {
+    mountedComponent(<CustomFieldSelectorDialog />, initialState);
+  });
+
+  it("should render the CustomFieldSelectorDialog component", () => {
+    expect(screen.getByRole("dialog")).toBeInTheDocument();
+  });
+
+  it("should render list of fields types", () => {
+    expect(screen.getByText("forms.type_label")).toBeInTheDocument();
+    expect(screen.getByText("forms.select_label")).toBeInTheDocument();
+    const elements = screen.getAllByTestId("field");
+    const totalCount = elements.length;
+
+    expect(totalCount).toBe(12);
+  });
+});
